Rename Filter/FilterPreview components
